+++ src/dune_rules/cinaps.ml
@@ -180,11 +180,13 @@ let gen_rules sctx t ~dir ~scope =
     Action.Full.make ~sandbox
     @@ A.chdir (Path.build dir)
          (A.progn
-            (A.run (Ok cinaps_exe) [ "-diff-cmd"; "-" ]
-            :: List.map cinapsed_files ~f:(fun fn ->
-                   A.diff ~optional:true (Path.build fn)
-                     (Path.Build.extend_basename fn ~suffix:".cinaps-corrected"))
-            ))
+            [ A.run (Ok cinaps_exe) [ "-diff-cmd"; "-" ]
+            ; A.concurrent
+              @@ List.map cinapsed_files ~f:(fun fn ->
+                     A.diff ~optional:true (Path.build fn)
+                       (Path.Build.extend_basename fn
+                          ~suffix:".cinaps-corrected"))
+            ])
   in
   let cinaps_alias =
     Alias.make ~dir @@ Option.value t.alias ~default:cinaps_alias